chore: styled-componentのFile構成の修正
