+++ hygen/component/new/component.tsx.t
@@ -4,12 +4,12 @@ unless_exists: true
 ---
 import React from 'react';
 
-import { Example } from './styled';
+import { Styled } from './styled';
 
 type Props = {
   name: string;
 };
 
 export const <%= name %>: React.FC<Props> = ({ name }) => {
-  return <Example>{name}</Example>;
+  return <Styled.Example>{name}</Styled.Example>;
 };
