+++ src/components/atoms/Text/styled.ts
@@ -1,0 +1,44 @@
+import { styled } from '@linaria/react';
+
+import { Color, colors } from '@/app/theme/colors';
+import { FontSize, fontSizes } from '@/app/theme/fontSizes';
+
+const lineHeights = {
+  normal: 1.2,
+  wide: 1.5,
+} as const;
+
+const fontWeights = {
+  normal: 400,
+  bold: 700,
+} as const;
+
+const letterSpacings = {
+  normal: '1px',
+  wide: '2px',
+} as const;
+
+const fontFamilies = {
+  normal: 'sans-serif',
+} as const;
+
+export type TextStyleProps = {
+  color: Color;
+  size: FontSize;
+  lineHeight?: keyof typeof lineHeights;
+  letterSpacing?: keyof typeof letterSpacings;
+  family?: keyof typeof fontFamilies;
+  weight?: keyof typeof fontWeights;
+};
+
+export const Styled = {
+  Text: styled.p<TextStyleProps>`
+    color: ${({ color }) => colors[color]};
+    font-size: ${({ size }) => fontSizes[size]};
+    font-weight: ${({ weight = 'normal' }) => fontWeights[weight]};
+    font-family: ${({ family = 'normal' }) => fontFamilies[family]};
+    line-height: ${({ lineHeight = 'normal' }) => lineHeights[lineHeight]};
+    letter-spacing: ${({ letterSpacing = 'normal' }) =>
+      letterSpacings[letterSpacing]};
+  `,
+};
